[ros2] Use unsigned int for track ID

+++ ros2/runner_cutter_control/include/runner_cutter_control/tracking/track.hpp
@@ -14,16 +14,16 @@ class Track {
     FAILED      // Failed to burn
   };
 
-  explicit Track(int id, std::pair<int, int> pixel,
+  explicit Track(uint32_t id, std::pair<int, int> pixel,
                  std::tuple<float, float, float> position,
                  State state = State::PENDING,
                  std::unique_ptr<Predictor> predictor = nullptr);
 
-  int getId() const { return id_; };
+  uint32_t getId() const { return id_; };
   std::pair<int, int> getPixel() const { return pixel_; };
   std::tuple<float, float, float> getPosition() const { return position_; };
   State getState() const { return state_; };
-  int getStateCount(State state) const;
+  size_t getStateCount(State state) const;
   Predictor& getPredictor() { return *predictor_; }
   const Predictor& getPredictor() const { return *predictor_; }
 
@@ -32,10 +32,10 @@ class Track {
   void setState(State state);
 
  private:
-  int id_{0};
+  uint32_t id_{0};
   std::pair<int, int> pixel_;
   std::tuple<float, float, float> position_;
-  std::unordered_map<State, int> stateCount_;
+  std::unordered_map<State, size_t> stateCount_;
   State state_{State::PENDING};
   std::unique_ptr<Predictor> predictor_;
 };
